notetaker-ai/requirements.txt
flask
flask-login
flask-sqlalchemy
openai
python-dotenv
werkzeug
pytest
pytest-flask